+++ CHANGES.rst
@@ -6,6 +6,7 @@ Features:
 
 Improvements:
  - Add RTL support
+ - PermalinkUtils is now able to parse a permalink
 
 Bugfix:
  - Fix crash when change visibility room (vector-im/riot-android#2679)